% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertRatio.R
\name{ConvertRatio}
\alias{ConvertRatio}
\title{Convert Ratios}
\usage{
ConvertRatio(
  year = "yr",
  mon = "mon",
  value.name = "value",
  data,
  to = "DJB",
  BaseTime = c(2013, 6)
)
}
\arguments{
\item{year}{a column name which represents a year.}

\item{mon}{a column name which represents a month.}

\item{value.name}{a column name, and its column reprensents level data
which are translated to year-on-year ratio, chain relative ratio or
fixed base ratio.}

\item{data}{a data.frame which include the year, mon and value.name columns at least.}

\item{to}{a charactor which represents a translated type. 'TB' means
a year-on-year ratio, 'HB' means a chain relative ratio, and 'DJB' means
a fixed base ratio.}

\item{BaseTime}{a numeric vector whose length is 2. When translated 'DJB',
it reprensents fixed base period. The 1st of \code{BaseTime} element means year,
and the 2nd element of \code{BaseTime} means month.}
}
\value{
Return a data frame the 1st column of which is year, the 2nd column of which
  is month, and the 3rd coloumn of which is the translated ratio.
}
\description{
Translation among year-on-year ratios, chain relative ratios and
  fixed base ratios.
}
\details{
\itemize{
   \item year-on-year ratio = \eqn{(V_{yr,mon}-V_{yr-1,mon})/V_{yr-1,mon}},
     where \eqn{V_{yr,mon}} denotes a level value in \code{mon} month \code{yr} year.
   \item chain relative ratio = \eqn{(V_t-V_{t-1})/V_{t-1}},
    where \eqn{V_t} denotes a level value at \eqn{t} period.
   \item fixed base ratio = \eqn{(V_t-V_{base})/V_{base}},
    where \eqn{V_{base}} denotes a level value at fixed base period.
 }
}
\examples{
data("profit")
profit[1:24,c('yr','mon','cumValue')]
##       yr mon cumValue
##    yr mon cumValue
##  2012   2  6060.06
##  2012   3 10449.14
##  2012   4 14525.20
##  2012   5 18434.13
##  2012   6 23116.61
##  2012   7 26784.50
##  2012   8 30596.66
##  2012   9 35239.53
##  2012  10 40240.05
##  2012  11 46625.26
##  2012  12 55577.70
##  2013   2  7091.53
##  2013   3 11740.08
##  2013   4 16106.91
##  2013   5 20812.45
##  2013   6 25836.64
##  2013   7 30032.16
##  2013   8 34863.90
##  2013   9 40452.81
##  2013  10 46263.22
##  2013  11 53338.01
##  2013  12 62831.02
# converte the cumValue coloumn to a year-on-year ratio,
ConvertRatio(year = 'yr',mon = 'mon',value.name = 'cumValue',data = profit,to = 'TB')
#converte the cumValue coloumn to a chain relative ratio ratio,
ConvertRatio(year = 'yr',mon = 'mon',value.name = 'cumValue',data = profit,to = 'HB')
# converte the cumValue coloumn to a fixed base ratio,
ConvertRatio(year = 'yr',mon = 'mon',value.name = 'cumValue',data = profit,to = 'DJB',
            BaseTime = c(2012,5))
}
