% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profit.R
\docType{data}
\name{profit}
\alias{profit}
\title{total profit of industrial enterprises in China}
\format{
A data frame with 88 rows and 7 variables:
\describe{
\item{date}{date}
\item{cumValue}{accumalative total profit, in current price RMB}
\item{cumTB}{year-on-year ratio for accumalative total profit}
\item{value}{total profit for the month, in current price RMB}
\item{TB}{year-on-year ratio for total profit for the month}
\item{yr}{year}
\item{mon}{month}
}
}
\source{
Wind Database
}
\usage{
profit
}
\description{
A dataset containing total profit of industrial enterprises in China.
  Time span is from 2012-2 to 2019-12.
}
\keyword{datasets}
